// generated/src/app/api-models/ar-tax-receipt.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ArTaxReceipt {
  seqCashReceipt: number;
  customerType: string;
  customerId: string;
  transDate: Date;
  transReceiptDate: Date;
  transactionType: string;
  companyCode: string;
  seqCashBatchId: number;
  glRefCode: string;
  taxPct: number;
  taxAmt: number;
  statementStatus: string;
  invoiceNo: number;
  postDate: Date;
  seqGpbilId: number;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  taxType: string;
}